Split EmHttpRequest on v1 and v2.
