+++ em-rest-client.gemspec
@@ -23,5 +23,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "bundler", "~> 1.9"
   spec.add_development_dependency "rake", "~> 10.0"
   spec.add_development_dependency "rspec", "~> 3.4.0"
-  spec.add_development_dependency "webmock", "~> 2.0"
+  spec.add_development_dependency "webmock", "~> 3.0"
 end
