+++ spec/em_http_request/execute/shared.rb
@@ -1,7 +1,13 @@
 shared_context :em_http_request_execute do
-  let(:args) { {} }
+  subject { request.execute }
 
   let(:request) { described_class.new(args) }
+  let(:args) { {} }
+  let(:accept_header) do
+    case RestClient.major_version
+    when 1 then '*/*; q=0.5, application/xml'
+    when 2 then '*/*'
+    end
+  end
 
-  subject { request.execute }
 end
